+++ test/missing.js
@@ -1,10 +1,9 @@
 // These are libraries that does not have any units test, they are
 // required here in order to track coverage for them.
 
-require('lib/clusterTools');
-require('lib/getInstances');
 require('lib/manager/instanceManager');
 require('lib/manager/modManager');
 require('lib/manager/pluginManager');
 require('lib/npmPostinstall.js');
 require('lib/typedef');
+require('../clusterctl.js');